feat: Add test for version record after author reply and review
A new test case has been added in 'src/docmap-parser.test.ts' to check the
version of record after the author reply and preprint review. The test
expects to match a certain object structure for the versioned reviewed
preprint.

Adjustments have also been made in 'src/test-fixtures/docmapGenerators.ts'
to facilitate the generation of the version of record and related
assertions.