feat: Refactor author response handling in docmap parser
This commit refactors the handling of author responses in the docmap parser. A new function 'addAuthorResponseToPreprint' has been introduced to handle the addition of author responses to the preprint. This function is now used in place of the previous method of adding author responses within the 'addEvaluationsToPreprint' function. The 'setPeerReviewFrom' function has been removed as it is no longer needed with this new approach.