fix: handle author response in same step as evaluations (#45)
* Add a failing test for reviewed preprint with a same step

* feat: Refactor author response handling in docmap parser

This commit refactors the handling of author responses in the docmap parser. A new function 'addAuthorResponseToPreprint' has been introduced to handle the addition of author responses to the preprint. This function is now used in place of the previous method of adding author responses within the 'addEvaluationsToPreprint' function. The 'setPeerReviewFrom' function has been removed as it is no longer needed with this new approach.

* feat: Improve handling of evaluations in preprint functions

This commit refactors the addAuthorResponseToPreprint and addEvaluationsToPreprint functions to directly pass the actions array instead of evaluations. This enables the functions to find and flatten all evaluations from the actions array internally, improving the abstraction and reducing redundancy in calling code.

---------

Co-authored-by: aisha <a.kapan@elifesciences.org>