feat: update docmap-parser to handle relatedContent in manuscript
This commit updates the docmap-parser in a way that it can now handle
the 'relatedContent' field in the manuscript data. If the field is present
in the found manuscript data, it is now added to the existing manuscript.
Additionally, a typo in a comment has been corrected (superceed to supersede).